apps/atlas-desktop/app-store/Blumtap/pyproject.toml
[tool.poetry]
name = "blum-tap-clicker"
version = "0.1.2"
description = "A fast python autoclicker is suitable for those who want to earn more Blum points"
authors = ["nhassl3 <thefanta284@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
django = "5.0.6"
djangorestframework = "^3.15.1"
django-cors-headers = "^4.3.1"
mss = "^9.0.1"
pyautogui = "^0.9.54"
opencv-python = "^4.10.0.82"
keyboard = "^0.13.5"
nuitka = "^2.3.4"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
